consolidated in `state` subdirectory: - useBoolean - useLink
